fixing the error in calculating the sum by changing the type of the variable from int to float.

+++ src/main/java/com/datastructures/maths/arithmeticprogression/model/ArithmeticResponse.java
@@ -7,6 +7,6 @@ public class ArithmeticResponse {
 	
 	private int outcomeCode;
 	private String outcomeMessage;
-	private int sumofArithmeticProg;
+	private float sumofArithmeticProg;
 
 }
